# Speed up new backtracking parser