Resolves #2727 
## Core Idea
The current parser; even when it is backtracking, builds nodes which is a terrible waste of time since every time we see a `match` or `case`, we need to copy the whole stack over and over in order to preserve the actual structure and not do any modifications on top of it. This is obviously not cheap. 

The following implementation does something very similar to old school state machines, where the parser is temporarily converted into a structure that tells you whether it accepts an input or not. It doesn't build any trees, at all. Which cuts a lot of time (see the benchmarks below). We simply change the parser's mode to accept-only from accept-and-build-tree when we are backtracking, and once we have the rule to follow, we switch back. 

The implementation is obviously not really straightforward, because lib2to3's code is very old and it is kind of hard to move pieces around. I did not want to make much modification to regular functions, in order to not cause any additional breakage. We simply change `push`/`pop`/`shift` from functions that create / modify nodes to functions that only modify the parser's state/transition table. This way, we don't need to copy the new nodes on each modification.

## Benchmarks

Benchmarking script is available [here](https://gist.github.com/isidentical/325325766daf7d3588b50704ef7bf064) (thanks to @ichard26 for sharing it with me).

> It'd be really appreciated if someone can also re-test these and share their results as well, for further verification.

An example that doesn't contain any PEP 634 syntax, but contains a lot of `match`/`case` keywords: [example.py](https://gist.github.com/isidentical/76d8b5e1777541d871ef1c4016952681).
```
-t py39 (main):: Mean +- std dev: 7.21 ms +- 0.25 ms
-t py39 (this branch):: Mean +- std dev: 7.25 ms +- 0.27 ms

-t py310 (main):: Mean +- std dev: 41.7 ms +- 2.5 ms
-t py310 (this branch):: Mean +- std dev: 13.8 ms +- 0.6 ms
```
From what I can interpret, this means that for `-t py310` there is a speed-up of ~**4X**.

An example that contains 25 non-PEP 634 usages and 25 PEP 634 usages of `match/case`: [example_2.py](https://gist.github.com/isidentical/0bfc74c21d351a87ae933087fd36304d)
```
-t py310 (main):: Mean +- std dev: 1.22 sec +- 0.03 sec
-t py310 (this branch):: Mean +- std dev: 93.0 ms +- 6.8 ms
```

And another example is parsing the [`Lib/test/test_patma.py`](https://github.com/python/cpython/blob/e6497fe698f6e87344501a68ffdea106eafcb257/Lib/test/test_patma.py), which has all of the pattern matching statement variations known to the humans (I've only run `lib2to3_parse()` once since it takes so much time on the `main`, so assume this is `+/- 20 seconds`):
```
main: 2 minutes, 33 seconds
this branch: 2.499 seconds
```

## Verification

Obviously, a change like this, at the core of the parser, should be verified though I do not have many examples of pattern matching myself. My script to parse every variation of pattern matching statements (originating through `Lib/test/test_patma.py`) says it can parse all of the examples. 

I've also taken the diff of the following command and compared it, which yielded no difference in formatting between main/this branch (speed-wise it is like 50X faster):
```console
$ black -tpy310 --fast --diff ~/projects/cpython/Lib/test/test_patma.py
```

Other than that, I've also tried formatting the whole stdlib (including tests) and there is also no relevant changes. If you have other ideas to verify this, that would be wonderful.