@ichard26 2021-12-26
Man, I wish GH-2725 was already merged so I could run diff-shades across this PR but I'll do it manually (assuming it supports the necessary features for such a stress test :crossed_fingers:),