@isidentical 2021-12-26
> Man, I wish GH-2725 was already merged so I could run diff-shades across this PR but I'll do it manually (assuming it supports the necessary features for such a stress test crossed_fingers). Also don't mind the primer or test uvloop failures, those are happening on main too and I haven't able to fix the crash yet.

Thanks a lot @ichard26. If you want, I can try to rebase on top of it but I have no idea whether there are anything that would block me to do so.