@ichard26 2021-12-26
It probably won't work as GitHub places restrictions on where workflows must be to trigger (IIRC it has to be on the default branch -- although I say that and I just realized I made that PR from my fork and not a branch upstream whoops).