@isidentical 2021-12-27
I've just ran the [syntax_test_suite](https://github.com/isidentical/syntax_test_suite) with this patch, and @ichard26 ran the [diff-shades](https://github.com/ichard26/diff-shades) which as far as I can tell neither yielded any changes so I am marking this PR as ready for review.