@isidentical 2022-01-10
Self-note: after talking with @JelleZijlstra, we decided to merge `_safe` function to the original definitions and guard the immutable implementations with a flag.